The screen should display contacts available on the user’s device phone (Ask for contacts permission if required by the platform).
The search bar should allow the user to search through the contact list and find the contacts matching the search query.

(Optional) The cancel button should clear the search bar.
(Optional) The floating action button should highlight the contact matching the name in blue and scroll to it.

Contacts data should be fetched from a flutter plugin written by you on the platform of your choice (Android or iOS).

The source code of the plugin should be shared with the test project source code.

The result of the test task can be sent through github, gitlab, bitbucket link or a zip file.
Note: Make sure the project compiles properly and that there’s no missing file.